requirements.txt
python-telegram-bot==21.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask==3.0.3
replicate==0.34.2
pytest==8.3.3
